chore: move BeaconConsensusEngineHandle to separate file
